fixed graphs
