+++ src/Components/Homepage.js
@@ -154,10 +154,11 @@ function Homepage () {
         <Paper elevation={10} className={classes.root} style={{backgroundColor:'#333333'}}>
         <Card className={classes.root} style={{backgroundColor:'#333333'}}>
           <CardContent >
-            <div className="canvas-container">
+            <div className="canvas-container" style={{width:'100%', height:'80%'}}>
             <Bar  backgroundColor="#333333"
             data={chartData1}
             options={{
+              responsive:true,
               title:{
               display:true,
               text:'Popularity',
